Integrate stakepools through the HTTP API (#216)
The purchase tickets view has been reworked slightly to add a menu to
choose between no stakepool, manual stakepool input, or use one of the
automatically configured pools. A new button has been added to manage
these pools. Clicking it opens a dialog window that lists all
available pools and allows the user to input their stakepool API key.

When one of the configured stakepools is selected, all of the fields
that must be manually specified for a pool are hidden.